backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.1
ecdsa==0.19.1
idna==3.11
pillow==12.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
PyJWT==2.10.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.4
urllib3==2.6.2
